requirements.txt
Django<=4.1
psycopg2>=2.8
ipdb<=0.14.0
djangorestframework<=3.14
drf-nested-routers==0.93.4
